refactor: move input compatibility wrapper
